Improve Performance for TriggerActionFlow
Improving the performance of TriggerActionFlow by invoking multiple Flows instead of one flow with  sObject record collection variables.
This decreases the execution time for processing an update on 200 cases from  approximately 15 seconds in the previous verion of `TriggerActionFlow` to about 3 seconds.

This is still a tax of approximately 2 seconds when compared to record processing time without any automation, but is a big improvement.

This change also makes the act of writing Trigger Action Flows more consistent with writing regular flows, since the flows now operate on one record at a time.